WIP refactor sign task
