+++ crates/operator/src/lib.rs
@@ -4,3 +4,5 @@
 pub mod client;
 /// Error
 pub mod error;
+/// Traits
+pub mod traits;
