+++ crates/operator_2_impl/src/builder.rs
@@ -140,20 +140,6 @@ impl OperatorBuilder {
         }
         Ok(())
     }
-
-    /// Sign the task response
-    pub fn sign_task_response(
-        &self,
-        task_response: TaskResponse,
-    ) -> Result<SignedTaskResponse, OperatorError> {
-        let encoded_response = TaskResponse::abi_encode(&task_response);
-        let hash_msg = keccak256(encoded_response);
-
-        let signed_msg = self.key_pair.sign_message(&hash_msg);
-        let signed_task_response =
-            SignedTaskResponse::new(task_response, signed_msg, self.operator_id);
-        Ok(signed_task_response)
-    }
 }
 
 #[cfg(test)]
